Added support for checking digital signatures (Test-DigitalSignature) and made some other updates
Added an example function (Test-DigitalSignature) that builds on
Get-AuthenticodeSignature on Windows 7 to also check if a file is
Catalog signed. This would be the case for Windows OS binaries, as well
as, some 3rd party binaries that are packaged up and signed together
(possibly PowerShell modules).

I also went through and made some updates to other functions. In
particular, GetTokenInformation is almost complete. This function has a
TOKEN_INFORMATION_CLASS argument that defines what token details are
being requested and returns those requested details. I believe I have
support for about 35 of the 41 options.

Here is a list of Functions, Enumerations, and Structures that have been
added or updated in this commit.

Added
Functions:
- kernel32\CreateFile
- wintrust\CryptCATAdminAcquireContext
- wintrust\CryptCATAdminCalcHashFromFileHandle
- wintrust\CryptCATAdminEnumCatalogFromHash
- wintrust\CryptCATAdminReleaseCatalogContext
- wintrust\CryptCATAdminReleaseContext
- wintrust\CryptCATCatalogInfoFromContext
- wintrust\WinVerifyTrust

Enumerations:
- SE_PRIVILEGE
- WTD_CHOICE
- WTD_FLAGS
- WTD_REVOKE
- WTD_STATEACTION
- WTD_UI
- WTD_UICONTEXT

Structures:
- CATALOG_INFO
- CLAIM_SECURITY_ATTRIBUTES_INFORMATION
- CLAIM_SECURITY_ATTRIBUTE_FQBN_VALUE
- CLAIM_SECURITY_ATTRIBUTE_OCTET_STRING_VALUE
- CLAIM_SECURITY_ATTRIBUTE_V1
- CTL_CONTEXT
- WINTRUST_BLOB_INFO
- WINTRUST_CATALOG_INFO
- WINTRUST_CERT_INFO
- WINTRUST_DATA
- WINTRUST_FILE_INFO
- WINTRUST_SGNR_INFO
- WINTRUST_SIGNATURE_SETTINGS

Updates
Functions:
- advapi32\ConvertSidToStringSid
- advapi32\DuplicateToken
- advapi32\GetTokenInformation
- iphlpapi\GetIpNetTable
- samlib\SamConnect
- wtsapi32\WTSQuerySessionInformation

Enumerations:
- MEMORY_STATE
- SE_GROUP
- WTS_CONNECTSTATE_CLASS

Structures:
- LUID_AND_ATTRIBUTES
- SID_AND_ATTRIBUTES
- SID_AND_ATTRIBUTES_HASH
- TOKEN_ACCESS_INFORMATION
- TOKEN_GROUPS
- TOKEN_GROUPS_AND_PRIVILEGES