Updated Get-DigitalSignature to use CreateFile instead of .NET to get File Handle
We were using [System.IO.File]::OpenRead() to get a handle to our target
file, but this was causing a"The system could not find the environment
option that was entered" error. I changed the code to use the CreateFile
function to retrieve the File Handle and have not run across the error
since.